fix the upgrade command to heed the -n option in `remove'
modified: scripts/install-cm3-compiler.sh

+++ scripts/install-cm3-compiler.sh
@@ -136,9 +136,11 @@ backup_old()
 
 rm_curent()
 {
-  rm -f "${FRONTEND}"
-  if [ "${GCC_BACKEND}" = yes ] ; then
-    rm -f "${BACKEND}"
+  if [ "${NOACTION}" != yes ] ; then
+    rm -f "${FRONTEND}"
+    if [ "${GCC_BACKEND}" = yes ] ; then
+      rm -f "${BACKEND}"
+    fi
   fi
 }
